# Improve the performance of -[NSObject valueForKeyPath:].