This replaces `_NSKVCSplitKeypath`, a key-splitting workhorse used in KVC and KVO, with a significantly more svelte version.

The old version:
* Used `dataUsingEncoding:` to take a full copy of the buffer at every step.
* Used ICU to find `.`

The new version:
* Trusts that CFString is capable of finding strings itself.

### Average per-invocation times.

```
  Old  |  New   | Workload
--------------------------
4208ns | 2762ns | "A.B.C.D" (2MM hits)
4985ns |  293ns | empty strings
4568ns | 1415ns | real-world profiling in application (56,000 hits @ string lengths 0-65)
```

Because both `valueForKey:` and `valueForKeyPath:` can be overridden by NSObject subclasses, full string copies of keypath components must be taken at every depth. This provides for a nearly-negligible performance impact (~799ns p/i in the average case.)

This pull request also fixes a regression in `valueForKey:` that caused the NSInvocation codepath to be triggered when the returned value would otherwise truly be `nil`.

#904.