@DHowett-MSFT 2016-09-15
There are tabs in one of the files.
I hate Visual Studio at times.