@DHowett-MSFT 2016-09-15
@bbowman your concerns regarding `type_encoding_cases` have been addressed.