@MSFTFox 2016-09-20
> //

nit: update copyright.

---
Refers to: Frameworks/Foundation/NSKeyValueCoding.mm:4 in f8418ea. [](commit_id = f8418ea3c0f1e7e1b5ccf625acd934351fa658e4, deletion_comment = False)